[ediciones] cambios presencial

+++ database/migrations/2024_05_23_021503_create_edicion_table.php
@@ -4,7 +4,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      *
@@ -19,7 +20,7 @@ return new class extends Migration {
             $table->string('fechaFinal')->nullable();
             $table->integer('idCampeon')->nullable();
             $table->boolean('liguilla')->default(false);
-            $table->integer('idCampeonato');
+            $table->bigInteger('idCampeonato');
             $table->foreign('idCampeonato')->references('id')->on('campeonatos');
         });
     }
